More optimizations by replacing int8/16 by int32

+++ include/internal/cv_ball_detector_seq.hpp
@@ -203,13 +203,13 @@ class BallDetector<TRIK_VIDTRANSCODE_CV_VIDEO_FORMAT_YUV422, TRIK_VIDTRANSCODE_C
       const int8_t* restrict srcImageRow = _inImage.m_ptr;
       uint64_t* restrict rgb2x888ptr = reinterpret_cast<uint64_t*>(s_rgb888);
       uint64_t* restrict hsvx2ptr    = reinterpret_cast<uint64_t*>(s_hsv);
-      const uint16_t height     = m_inImageDesc.m_height;
+      const uint32_t height     = m_inImageDesc.m_height;
       const uint32_t lineLength = m_inImageDesc.m_lineLength;
-      const uint16_t widthxu16  = m_inImageDesc.m_width*sizeof(uint16_t);
+      const uint32_t widthxu16  = m_inImageDesc.m_width*sizeof(uint16_t);
 
       assert(m_inImageDesc.m_height % 4 == 0); // verified in setup
 #pragma MUST_ITERATE(4, ,4)
-      for (uint16_t heightRemains=height; heightRemains>0; --heightRemains)
+      for (uint32_t heightRemains=height; heightRemains>0; --heightRemains)
       {
         assert(reinterpret_cast<intptr_t>(srcImageRow) % 8 == 0); // let's pray...
         const uint64_t* restrict srcImagex2 = reinterpret_cast<const uint64_t*>(srcImageRow);
